shan-associates/backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
djongo==1.3.6
pymongo==4.4.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.2.2